chore(firstMile): move python steps into steps/python directory
