+++ src/homepageExperience/components/steps/python/ExecuteAggregateQuery.tsx
